split-out io functions
